Refactor admin/client pages to use FC and typescript
